+++ lms/djangoapps/notification_prefs/views.py
@@ -10,7 +10,7 @@ from django.core.exceptions import PermissionDenied
 from django.http import Http404, HttpResponse
 from django.views.decorators.http import require_GET, require_POST
 
-from mitxmako.shortcuts import render_to_response
+from edxmako.shortcuts import render_to_response
 from notification_prefs import NOTIFICATION_PREF_KEY
 from user_api.models import UserPreference
 
@@ -153,12 +153,14 @@ def ajax_status(request):
 
 
 @require_GET
-def unsubscribe(request, token):
+def set_subscription(request, token, subscribe):  # pylint: disable=unused-argument
     """
-    A view that disables notifications for a user who may not be authenticated
+    A view that disables or re-enables notifications for a user who may not be authenticated
 
     This view is meant to be the target of an unsubscribe link. The request
     must be a GET, and the `token` parameter must decrypt to a valid username.
+    The subscribe flag feature controls whether the view subscribes or unsubscribes the user, with subscribe=True
+    used to "undo" accidentally clicking on the unsubscribe link
 
     A 405 will be returned if the request method is not GET. A 404 will be
     returned if the token parameter does not decrypt to a valid username. On
@@ -174,6 +176,13 @@ def unsubscribe(request, token):
     except User.DoesNotExist:
         raise Http404("username")
 
-    UserPreference.objects.filter(user=user, key=NOTIFICATION_PREF_KEY).delete()
-
-    return render_to_response("unsubscribe.html", {})
+    if subscribe:
+        UserPreference.objects.get_or_create(user=user,
+                                             key=NOTIFICATION_PREF_KEY,
+                                             defaults={
+                                                 "value": UsernameCipher.encrypt(user.username)
+                                             })
+        return render_to_response("resubscribe.html", {'token': token})
+    else:
+        UserPreference.objects.filter(user=user, key=NOTIFICATION_PREF_KEY).delete()
+        return render_to_response("unsubscribe.html", {'token': token})
